requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
python-dotenv==0.10.3
SQLAlchemy==1.3.10
WTForms==2.2.1
PyJWT==1.7.1
Flask-Mail==0.9.1
Click>=7.0
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.9.0
requests==2.22.0
python-dateutil>=2.8.1
pyenchant>=2.0.0
guess-language-spirit>=0.5.3